Refactor import and call to dom
Refactor import 'processWeatherData' to 'extractRelevantData' to reflect
changes in the process-weather-data.js module. In the queryWeatherAPI fn,
rather than extracting the data and returning to this module to then
send the final object to the dom, just send the object via the
extractRelevantData fn to the process-weather-data.js fn. The object is
then transferred to the dom from there